Rest_Api/pyproject.toml
[tool.poetry]
name = "HW_14"
version = "0.1.0"
description = "HW_14"
authors = ["Valekantina <vgordynska@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.22.0"
fastapi = "^0.95.1"
SQLAlchemy = "^2.0.11"
psycopg2 = "^2.9.6"
alembic = "^1.10.4"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"
fastapi-mail = "^1.2.8"
fastapi-limiter = "^0.1.5"
python-dotenv = "^1.0.0"
redis = "^4.5.5"
cloudinary = "^1.32.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.0.1"
httpx = "^0.24.1"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]